Minor type adjustment

+++ tpot_exported_pipeline.py
@@ -4,7 +4,7 @@ from sklearn.linear_model import LogisticRegression
 from sklearn.model_selection import train_test_split
 
 # NOTE: Make sure that the class is labeled 'target' in the data file
-tpot_data = pd.read_csv('/Users/minhanho/Documents/BCB330/TF_FeatureExtraction/features_with_level1class.csv',sep='COLUMN_SEPARATOR', converters={'cell_id': float})
+tpot_data = pd.read_csv('/Users/minhanho/Documents/BCB330/TF_FeatureExtraction/features_with_level1class.csv',sep='COLUMN_SEPARATOR', converters={'cell_id': np.float64})
 features = tpot_data.drop('target', axis=1).values
 training_features, testing_features, training_target, testing_target = \
             train_test_split(features, tpot_data['target'].values, random_state=42)
